+++ Assets/Banechi/StageIDHolder.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d4d128ed384cbb04a96b9fae44f1ccec
+guid: 15a847bd39bb312449a20a774d10c5e0
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
